+++ react/__mocks__/vtex.render-runtime.ts
@@ -1,2 +1,3 @@
 export const global = {}
 export const Window = {}
+export const canUseDOM = true