+++ react/components/HighlightOverlay/hooks/useHighlightedElementInfo.ts
@@ -19,7 +19,6 @@ function getElementInfo(
   )
 
   const provider = document.querySelector<HTMLDivElement>('.render-provider')
-
   const elementsArray: Element[] = Array.prototype.slice.call(elements)
 
   const viewportMiddle =
@@ -77,7 +76,8 @@ function getElementInfo(
 
 export default function useHighlightedElementInfo(
   highlightTreePath: State['highlightTreePath'],
-  sidebarBlocksMap: State['sidebarBlocksMap']
+  sidebarBlocksMap: State['sidebarBlocksMap'],
+  elementHeight: State['elementHeight']
 ) {
   return useMemo<Partial<ReturnType<typeof getElementInfo>>>(() => {
     return (
@@ -85,5 +85,6 @@ export default function useHighlightedElementInfo(
         getElementInfo(highlightTreePath, sidebarBlocksMap)) ||
       {}
     )
-  }, [highlightTreePath, sidebarBlocksMap])
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [highlightTreePath, sidebarBlocksMap, elementHeight])
 }