+++ react/components/HighlightOverlay/index.tsx
@@ -2,11 +2,13 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import { CSSTransition } from 'react-transition-group'
 import { canUseDOM } from 'vtex.render-runtime'
+
 import useAutoScroll from './hooks/useAutoScroll'
 import useHighlightOnHover from './hooks/useHighlightOnHover'
-import usePortal from './hooks/usePortal'
-import useStyles from './hooks/useStyles'
 import useHighlightedElementInfo from './hooks/useHighlightedElementInfo'
+import usePortal from './hooks/usePortal'
+import useResizeObserver from './hooks/useResizeObserver'
+import useStyles from './hooks/useStyles'
 import OverlayMask from './OverlayMask'
 import { State } from './typings'
 
@@ -45,7 +47,9 @@ const HighlightOverlay: React.FC<Props> = props => {
     sidebarBlocksMap,
   } = state
 
-  const portalContainer = usePortal()
+  const { subscribeToResize, unsubscribeToResize } = useResizeObserver()
+
+  const portalContainer = usePortal({ subscribeToResize, unsubscribeToResize })
 
   useHighlightOnHover(state)
 
@@ -74,14 +78,23 @@ const HighlightOverlay: React.FC<Props> = props => {
 
   const { visibleElement, hasValidElement } = useHighlightedElementInfo(
     titleTreePath,
-    sidebarBlocksMap
+    sidebarBlocksMap,
+    state.elementHeight
   ) || { visibleElement: undefined, hasValidElement: false }
 
   useAutoScroll({ editMode, highlightTreePath, visibleElement })
 
-  const { highlightStyle, labelStyle, maskStyle } = useStyles({
+  const isOverlayMaskActive = Boolean(
+    hasValidElement && openBlockTreePath && portalContainer
+  )
+
+  useStyles({
     hasValidElement,
     highlightTreePath: titleTreePath,
+    isOverlayMaskActive,
+    setState,
+    subscribeToResize,
+    unsubscribeToResize,
     visibleElement,
   })
 
@@ -104,14 +117,15 @@ const HighlightOverlay: React.FC<Props> = props => {
         timeout={150}
       >
         <div
+          data-testid="editor-provider-overlay"
           id="editor-provider-overlay"
-          style={highlightStyle}
+          style={state.highlightStyle}
           className="absolute bw2 ba"
         >
           {title && (
             <p
               className="absolute white f7 ma0 right-0 ph2 pb2 pt1 truncate tc"
-              style={labelStyle}
+              style={state.labelStyle}
               title={title}
             >
               {title}
@@ -120,12 +134,7 @@ const HighlightOverlay: React.FC<Props> = props => {
         </div>
       </CSSTransition>
       {ReactDOM.createPortal(
-        <OverlayMask
-          style={maskStyle}
-          isActive={Boolean(
-            hasValidElement && openBlockTreePath && portalContainer
-          )}
-        />,
+        <OverlayMask style={state.maskStyle} isActive={isOverlayMaskActive} />,
         portalContainer
       )}
     </>